server/requirements.txt
# Python dependencies for the FastAPI server
fastapi
uvicorn
python-multipart
google-generativeai
pdf2image
pytesseract
Pillow
python-dotenv